requirements.txt
flask
transformers
sentence-transformers
python-docx
PyMuPDF
pdfminer.six
